Rename cli, add method to open url automatically
