+++ lua/gitportal/core.lua
@@ -1,9 +1,10 @@
 local ui = require('gitportal.ui')
+local cli = require('gitportal.cli')
 local git_helpers = require('gitportal.git')
 
 
 vim.api.nvim_create_user_command('Gplink', function()
-  ui.open_window(git_helpers.get_git_url_for_current_file())
+  cli.open_link_in_browser(git_helpers.get_git_url_for_current_file())
 end, {})
